+++ scripts/start.py
@@ -106,7 +106,8 @@ def main_menu():
                         current_level[1] = 0
                         with open(CUR_LEVEL_PATH, 'w') as file:
                             file.write(','.join(map(str, current_level[:2])))
-                        os.remove(CUR_LEVEL_PROGRESS_PATH)
+                        if os.path.exists(CUR_LEVEL_PROGRESS_PATH):
+                            os.remove(CUR_LEVEL_PROGRESS_PATH)
                 if level_selection:
                     clicked_level = levels.get_click(mouse_pos)
                     if clicked_level is not None:
